Check if a provider is used before checking if the subdomain is vulnerable
There is still a setup where a subdomain points to a vulnerable service's IP (not CNAME) and you can take it over (https://youtu.be/HhJv8CU-RIk?t=1396).
So we may want to explore how we can check for those, but I think this setup is so rare it doesn't warrant a fix right now.